requirements.txt
streamlit
pyrebase4
secure-smtplib
nltk
requests
bs4 
selenium
webdriver-manager
youtube-comment-downloader